requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
openai==0.27.8
Flask-Migrate==4.0.5
psycopg2==2.9.9
python-dotenv==1.0.0
anthropic==0.19.1